Return uint8_t * from marchingCubes

+++ march.h
@@ -1,6 +1,8 @@
 #ifndef MARCH_H
 #define MARCH_H
 
-float* marchingCubes(int dims[3], float *potential, float shift[3], float scale[3]);
+#include <cstdint>
+
+uint8_t* marchingCubes(int dims[3], float *potential, float shift[3], float scale[3]);
 
 #endif
